+++ modules/worker_manager/tests/test_worker_property_data.py
@@ -2,7 +2,7 @@
 Test queue property.
 """
 
-from modules.worker_manager import worker_manager
+from modules.worker_manager import worker_property_data
 
 
 # Test functions use test fixture signature names and access class privates
@@ -27,7 +27,7 @@ class TestCreate:
         """
         count = 5
 
-        result, worker_property = worker_manager.WorkerPropertyData.create(count, stub, (), [], [])
+        result, worker_property = worker_property_data.WorkerPropertyData.create(count, stub, (), [], [])
 
         assert result
         assert worker_property is not None
@@ -38,7 +38,7 @@ class TestCreate:
         """
         count = 0
 
-        result, worker_property = worker_manager.WorkerPropertyData.create(count, stub, (), [], [])
+        result, worker_property = worker_property_data.WorkerPropertyData.create(count, stub, (), [], [])
 
         assert not result
         assert worker_property is None
@@ -49,7 +49,7 @@ class TestCreate:
         """
         count = -1
 
-        result, worker_property = worker_manager.WorkerPropertyData.create(count, stub, (), [], [])
+        result, worker_property = worker_property_data.WorkerPropertyData.create(count, stub, (), [], [])
 
         assert not result
         assert worker_property is None
